fixed count algorithm
